Setting `sqlalchemy` restrictions to be compatible with our version of `pandas` (#1069)
This PR adds a runtime restriction on `sqlalchemy<2.0` because version 2.0 is incompatible with `pandas=1.3`. See [here](https://stackoverflow.com/a/75282604) for more information.

The runtime restriction was missing even though it was included in the conda dev dependencies.

Included in this PR are changes to remove Git LFS from the base image and only install it in later stages used to build the conda package.

Authors:
  - Michael Demoret (https://github.com/mdemoret-nv)

Approvers:
  - David Gardner (https://github.com/dagardner-nv)

URL: https://github.com/nv-morpheus/Morpheus/pull/1069