Resolve camouflage statup issues by increasing timeout to 30 seconds.
Improve camouflage error reporting by redirecting stdout and stderr to a file.
This is a work-around for not being able to control the log level of camouflage's console logger
https://github.com/testinggospels/camouflage/issues/244

We can't redirect stdout/stderr to a PIPE since on successful startup the PIPE would fill and block
camouflage. On the other hand sending stdout/stderr to /dev/null prevents users from seeing errors
from the shell/OS regarding camouflage failing to start.