Revert the relative path change used in examples (#1978)
This change was made because of check-style failures when tests were refactored but is causing import failures.

Note: This commit also reverts the style fix and the local imports are mis-sorted. We may need to update the style checker to fix it post release.

Authors:
  - Anuradha Karuppiah (https://github.com/AnuradhaKaruppiah)

Approvers:
  - Yuchen Zhang (https://github.com/yczhang-nv)

URL: https://github.com/nv-morpheus/Morpheus/pull/1978